feat: enable sending of files other than stickers
idea is that if a bot receives a supported file type, it forwards to a private file channel containing both angel and mortal bot.
then, the other bot (that didn’t forward the message) receives and uses the file id from the forwarded message to send the file to the angel/mortal. i
the forwarded message is then deleted from the file channel by the bot.
currently unsupported file types:
* location: live location is updated as many separate messages, does not seem well supported for bots
* contacts: no need to provide contacts since no one sends contacts, can easily send a screenshot or type the phone number out
* game: only a few games that can be played currently, not significant

other minor changes:
* removed previous comments for file sending
* printing timestamp for every message received, instead of just text messages